chatbot-back/requirements.txt
Flask
google-cloud-automl